test: Add linting of Python test scripts
We currently have one test moved to the new exeter based framwork written
in Python.  We plan to add many more, so add linting (flake8) and type
checking (mypy) of those scripts.  This can be invoked manually with
"make flake8" or "make mypy" in test/, and is also added to the static
checkers test set.

Signed-off-by: David Gibson <david@gibson.dropbear.id.au>
Signed-off-by: Stefano Brivio <sbrivio@redhat.com>