tcp, udp: Don't exclude ports in {tcp,udp}_port_rebind()
To avoid circular forwarding, {tcp,udp}_port_rebind() refuse to listen on
ports that we're already listening on in the reverse direction.  This is
redundant, because we already remove such ports from the forward map when
we scan.  This was needed previously, because our reverse maps might have
been one cycle out of date, so could be missing a newly appeared port.

We've now rearranged the port scanning code to avoid that, so we don't need
the check in tcp_port_rebind() any more.

Signed-off-by: David Gibson <david@gibson.dropbear.id.au>
Signed-off-by: Stefano Brivio <sbrivio@redhat.com>