util: Correct error message on SO_BINDTODEVICE failure
The warning message we print of SO_BINDTODEVICE fails is incorrect: we
include EPOLL_TYPE_STR(proto), but the proto variable is not an epoll type
so this will generate misleading garbage.  This has been wrong ever since
340164445 but the message is rare enough that we never noticed.

Correct that, and reword the message a bit for clarity while we're
there.

Fixes: 340164445341 ("epoll: Generalize epoll_ref to cover things other than sockets")
Signed-off-by: David Gibson <david@gibson.dropbear.id.au>
Signed-off-by: Stefano Brivio <sbrivio@redhat.com>